route: Improved performance
This commit improves performance in the following two places:
1. In the cleanNodes function, the extra if case reduces the number
of writes, which improves performance. The effect is strongest when
multiple instances are running at the same time.
2. Eliminated some calls to putchar by adding a comma to the
JSON title.

Future areas of potential performance improvement include:
Using a different cleanNodes function when the route is very simple
Using an output buffer to reduce the number of libc calls
Mmapping the title file