buffer: Add new bufferExpand and bufferAddUnsafe functions
bufferExpand can be used to ensure a minimal amount of free space left.
This can be used to reduce the number of required reallocations when
performing some bufferAdds later.
If enough free space is allocated using bufferExpand it is possible to
use bufferAddUnsafe, this function does not cehck whether it needs to
realloc the memory and should therefore be slightly faster in some cases.