route: Performance optimizations
This commit changes the m ment when the distBs of nodes are marked.
Now, the distBs of the references are marked during reference copying.
Deduplication is now also done during reference cpying.

Doing this optimization is a bit more complicated for distA, because
for those we also have a list containing all of the nodes that have
been visited (toCleanA). Therefore this is not included in this commit.

This commit also changes the statistics and can also change the values of
dist.a and dist.b. The total distance should still be correct.

There might still be a few small bugs in this commit.