route: Improved performance
In total this series of commits improves the performance by approx
a factor 2. The output is the same, except for the a and b parts of
the dist subobject. The dist.a+dist.b is still the same.

These improvements have also made some off the stats less meaningfull.
Further commits will address that issue.