route: Add option to exclude certain articles
This works by marking things to be excluded using a
special valuw in the dists array and then ignoring
nodes that have that specific value.

The change in performance of route is less than 2%
for routes that don't use exclude. It might be
possoble to make it a bit faster for routes with
exclude by placing the exclude check in the innermost loop.