parsepagelinks2: Support new pagelinks.sql format
This new format has the linktarget id instead of a string. Therefore
we must use the linktarget.sql file to convert this id to the
corresponding string.

parsepagelinks2 now also parses page.sql and linktarget.sql itself,
instead of using the result from another program. Therefore parsepagetitles
is no longer needed.

The parsing of page.sql and redirect.sql could be multithreaded, but
it isn't right now, because it only saves 20 seconds at most.

Rough worlkflow for parsing the sql dumps with this program:

    zcat pagelinks.sql.gz | ./parsepagelinks | sort -n >pagelinks.txt
    ./parsepagelinks2 <(zcat page.sql.gz) <(zcat linktarget.sql.gz) pagelinks.txt <(zcat redirect.sql.gz) | tr '_' ' ' | ./presort | LC_ALL=C sort | cut -b 2- | ./removeduplicates >links.txt
    ./parselinks links.txt nodes.bin titles.txt